refactor(code-drawing): enhance CodeDrawingElement with new rendering logic and toolbar
- Introduced a custom hook for managing code drawing element state, including debounced rendering and error handling.
- Replaced BaseCodeDrawingElement with PlateElement for improved integration.
- Added a toolbar for selecting drawing type and view mode, enhancing user interaction.
- Refactored CodeDrawingPreview and CodeDrawingTextarea components for better state management and responsiveness.
