--- apps/www/src/registry/ui/code-drawing-preview-area.tsx
@@ -1,73 +1,0 @@
-'use client';
-
-import * as React from 'react';
-
-import { VIEW_MODE, type ViewMode } from '@platejs/code-drawing';
-
-export function CodeDrawingPreviewArea({
-  image,
-  loading,
-  code,
-  viewMode,
-  readOnly: _readOnly = false,
-  isMobile = false,
-  showBorder = false,
-  toolbar,
-}: {
-  image: string;
-  loading: boolean;
-  code: string;
-  viewMode: ViewMode;
-  readOnly?: boolean;
-  isMobile?: boolean;
-  showBorder?: boolean;
-  toolbar?: React.ReactNode;
-}) {
-  const showImage = viewMode === VIEW_MODE.Both || viewMode === VIEW_MODE.Image;
-
-  return (
-    <div
-      className={`flex min-w-0 flex-1 flex-col ${isMobile ? '' : 'relative'} ${
-        showBorder && isMobile ? 'border-b' : ''
-      }`}
-    >
-      {toolbar && (
-        <div
-          className={
-            isMobile
-              ? 'mt-2 mb-2 flex justify-end px-2'
-              : 'absolute right-2 z-10 mt-2'
-          }
-        >
-          {toolbar}
-        </div>
-      )}
-
-      {showImage ? (
-        <div
-          className={
-            'flex flex-1 items-center justify-center rounded-md bg-muted/30 p-4'
-          }
-        >
-          {loading && <div className="text-muted-foreground">Loading...</div>}
-          {!loading && image && (
-            <img
-              src={image}
-              alt="Code drawing"
-              className="max-h-full max-w-full object-contain"
-            />
-          )}
-          {!loading && !image && (
-            <div className="text-muted-foreground">
-              {code.trim() ? 'Rendering...' : 'Preview will appear here'}
-            </div>
-          )}
-        </div>
-      ) : (
-        <div className="pointer-events-none flex flex-1 items-center justify-center rounded-md border bg-muted/30 p-4 opacity-0">
-          {/* Placeholder to maintain height */}
-        </div>
-      )}
-    </div>
-  );
-}
