--- packages/code-drawing/src/react/hooks/index.ts
@@ -1,5 +1,0 @@
-/**
- * @file Automatically generated by barrelsby.
- */
-
-export * from './useCodeDrawingElement';
