+++ packages/code-drawing/src/react/index.ts
@@ -2,6 +2,4 @@
  * @file Automatically generated by barrelsby.
  */
 
-export * from './CodeDrawingElement';
 export * from './CodeDrawingPlugin';
-export * from './hooks/index';
